[Bronze V] Title: 꼬마 정민, Time: 176 ms, Memory: 17660 KB - DashHub

+++ BOJ/Bronze/11382. 꼬마 정민/Main.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int res = sc.nextInt();
-        res += sc.nextInt();
-        res += sc.nextInt();
+        long res = sc.nextLong();
+        res += sc.nextLong();
+        res += sc.nextLong();
         System.out.println(res);
 
     }
